fix: tidy up codes
